se unifican controladores de productos
