--- controllers/productCrud.controller.js
@@ -1,7 +1,0 @@
-const productCrudCTRL = {
-    productCrud: (req, res)=>{
-        res.render('productCrud')
-    }
-};
-
-module.exports = productCrudCTRL;
