+++ controllers/products.controller.js
@@ -1,7 +1,24 @@
-const productDetails = {
-  details: (req, res) => {
-    res.render("productDetail");
-  },
-};
+const express = require('express');
+const router = express.Router();
+const fs = require('fs');
+const path = require('path');
 
-module.exports = productDetails;
+const productsFilePath = path.join(__dirname, '../data/products.json');
+const products = JSON.parse(fs.readFileSync(productsFilePath, 'utf-8'));
+
+const productsCtrl = {
+   
+    detail: (req, res) => {
+      idProd = req.params.id;
+		  producto = products.find(function (product) {
+			return product.id == idProd;
+		});
+      res.render("prodDetail",{producto});
+    },
+
+    productCrud: (req, res)=>{
+      res.render('productCrud')
+    }
+  };
+  
+  module.exports = productsCtrl;
