+++ index.js
@@ -5,9 +5,7 @@ const path = require("path");
 const indexController = require("./controllers/index.controller");
 const cartController = require("./controllers/cart.controller");
 const userCTRL = require("./controllers/user.controller");
-const productDetailCTRL = require("./controllers/productDetail.controller");
-const productCrudCTRL = require("./controllers/productCrud.controller");
-//const productsDetails = require("./controllers/products.controller");
+const productsCTRL = require("./controllers/products.controller");
 
 app.set("view engine", "ejs");
 app.set("views", "./views");
@@ -17,29 +15,11 @@ app.use("//", indexController.index);
 app.use("/cart", cartController.cart);
 app.use("/login", userCTRL.login);
 app.use("/register", userCTRL.register);
-app.use("/prodDetail/:id", productDetailCTRL.detail);
-app.use("/productCrud", productCrudCTRL.productCrud);
+app.use("/prodDetail/:id", productsCTRL.detail);
+app.use("/productCrud", productsCTRL.productCrud);
 
 
 app.listen(3000, () => {
   console.log("el servidor inicio...");
 });
 
-// app.get("/", (req, res) => {
-//   res.sendFile(path.join(__dirname, "views/index.html"));
-// });
-
-// app.get("/prod-detail", (req, res) => {
-//   res.sendFile(path.join(__dirname, "views/productDetail.html"));
-// });
-
-// /*ruta parametrizada de prueba Gerardo*/
-// app.get("/productos/:id", function (req, res) {
-//   let idProducto = req.params.id;
-//   res.sendFile(
-//     path.join(__dirname, "views/productDetail-" + idProducto + ".html")
-//   );
-// });
-// /*fin ruta parametrizada de prueba para detalles de producto*/
-
-// //app.use("/prodDetail", prodDetailCtrl.load);
